move security hub finding interface to common
